% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{add_ohlcv_transformations}
\alias{add_ohlcv_transformations}
\title{Add OHLCV Data Transformations}
\usage{
add_ohlcv_transformations(dt)
}
\arguments{
\item{dt}{data.table with OHLCV data}
}
\value{
data.table with additional transformations
}
\description{
Adds efficient data transformations using data.table
}
