% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{aggregate_ohlcv}
\alias{aggregate_ohlcv}
\title{Efficient Data Aggregation}
\usage{
aggregate_ohlcv(dt, period = "weekly")
}
\arguments{
\item{dt}{data.table with OHLCV data}

\item{period}{Aggregation period ("daily", "weekly", "monthly")}
}
\value{
Aggregated data.table
}
\description{
Fast data aggregation using data.table
}
