% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{cache_data}
\alias{cache_data}
\title{Cache Data}
\usage{
cache_data(data, symbol, cache_config = list())
}
\arguments{
\item{data}{Data to cache}

\item{symbol}{Symbol name}

\item{cache_config}{Cache configuration}
}
\value{
Logical indicating success
}
\description{
Simple file-based caching for MVP
}
