% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{detect_anomalies}
\alias{detect_anomalies}
\title{Detect Data Anomalies}
\usage{
detect_anomalies(data, method = "statistical")
}
\arguments{
\item{data}{Financial data}

\item{method}{Detection method}
}
\value{
List of detected anomalies
}
\description{
Detects anomalies in financial data
}
