% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{fetch_ohlcv.AdvancedProviderManager}
\alias{fetch_ohlcv.AdvancedProviderManager}
\title{Fetch OHLCV with Advanced Failover}
\usage{
\method{fetch_ohlcv}{AdvancedProviderManager}(provider, symbol, start_date, end_date, ...)
}
\arguments{
\item{provider}{AdvancedProviderManager object}

\item{symbol}{Character symbol to fetch}

\item{start_date}{Start date for data}

\item{end_date}{End date for data}

\item{...}{Additional arguments}
}
\value{
data.table with OHLCV data
}
\description{
Fetch OHLCV with Advanced Failover
}
