% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{get_cached_data}
\alias{get_cached_data}
\title{Get Cached Data}
\usage{
get_cached_data(provider_manager, symbol, start_date, end_date)
}
\arguments{
\item{provider_manager}{AdvancedProviderManager object}

\item{symbol}{Symbol to fetch}

\item{start_date}{Start date}

\item{end_date}{End date}
}
\value{
Cached data or NULL
}
\description{
Get Cached Data
}
