% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_provider.R
\name{get_ib_contract_details}
\alias{get_ib_contract_details}
\title{Get IB Contract Details}
\usage{
get_ib_contract_details(provider, symbol)
}
\arguments{
\item{provider}{IBProvider object}

\item{symbol}{Symbol to get details for}
}
\value{
List with contract details
}
\description{
Stub function for getting contract details from IB
}
