% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{get_provider_by_name}
\alias{get_provider_by_name}
\title{Get Provider by Name}
\usage{
get_provider_by_name(provider_manager, provider_name)
}
\arguments{
\item{provider_manager}{AdvancedProviderManager object}

\item{provider_name}{Name of provider to retrieve}
}
\value{
Provider object or NULL
}
\description{
Get Provider by Name
}
