% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{get_provider_order}
\alias{get_provider_order}
\title{Get Provider Order Based on Strategy}
\usage{
get_provider_order(provider_manager)
}
\arguments{
\item{provider_manager}{AdvancedProviderManager object}
}
\value{
Character vector of provider names in order
}
\description{
Get Provider Order Based on Strategy
}
