% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_provider.R
\name{is_ib_connected}
\alias{is_ib_connected}
\title{Check IB Connection Status}
\usage{
is_ib_connected(provider)
}
\arguments{
\item{provider}{IBProvider object}
}
\value{
Logical indicating if connected
}
\description{
Check IB Connection Status
}
