% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{ProviderManager}
\alias{ProviderManager}
\title{Provider Manager Class}
\usage{
ProviderManager(
  providers = list(),
  failover_strategy = character(0),
  circuit_breaker = NULL
)
}
\arguments{
\item{providers}{List of data provider objects}

\item{failover_strategy}{Character strategy for failover}

\item{circuit_breaker}{Circuit breaker configuration}
}
\description{
Manages multiple data providers with failover
}
