% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{record_response_time}
\alias{record_response_time}
\title{Record Response Time}
\usage{
record_response_time(provider_manager, provider_name, response_time)
}
\arguments{
\item{provider_manager}{AdvancedProviderManager object}

\item{provider_name}{Name of provider}

\item{response_time}{Response time in seconds}
}
\description{
Record Response Time
}
