% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_manager.R
\name{update_provider_health}
\alias{update_provider_health}
\title{Update Provider Health Metrics}
\usage{
update_provider_health(provider_manager, provider_name, success)
}
\arguments{
\item{provider_manager}{AdvancedProviderManager object}

\item{provider_name}{Name of provider}

\item{success}{Logical indicating success or failure}
}
\description{
Update Provider Health Metrics
}
