% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provider.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data Quality}
\usage{
validate_data(data, symbol, validation_rules = list())
}
\arguments{
\item{data}{Data to validate}

\item{symbol}{Symbol name}

\item{validation_rules}{List of validation rules}
}
\value{
DataValidation object
}
\description{
Validates fetched financial data
}
