% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgeroot-package.R
\name{fetch_ohlcv}
\alias{fetch_ohlcv}
\title{Fetch OHLCV Data}
\usage{
fetch_ohlcv(provider, symbol, start_date, end_date, ...)
}
\arguments{
\item{provider}{Provider object}

\item{symbol}{Character symbol to fetch}

\item{start_date}{Start date for data}

\item{end_date}{End date for data}

\item{...}{Additional arguments}
}
\value{
data.table with OHLCV data
}
\description{
Generic function to fetch OHLCV data from providers
}
